feat(admin): add navigation manager
